Product price float

+++ db/migrate/20130416134413_add_price_to_product.rb
@@ -1,5 +1,5 @@
 class AddPriceToProduct < ActiveRecord::Migration
   def change
-    add_column :products, :price, :double
+    add_column :products, :price, :float
   end
 end
